remove bug

+++ src/Settings.h
@@ -10,7 +10,7 @@ struct Settings{
     std::array<std::pair<short,short>,2> displayResolution = {{{1280,720}}}; //probably more in the future
     std::array<std::string,3> level = {"easy","medium","hard"};
     sf::Vector2f shipSize{50,50};
-    sf::Vector2f shipPosition{displayResolution[0].first/2, displayResolution[0].second-shipSize.y-shipSize.y};
+    sf::Vector2f shipPosition{static_cast<float>(displayResolution[0].first/2), displayResolution[0].second-shipSize.y-shipSize.y};
     sf::Vector2f normalBulletSize{3,3}; //BULLET SIZE
     sf::Vector2f curlyBulletSize{10,10};
     sf::Vector2f laserSize{5,2000};
